# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartRest AIoT API",
        "description": "RESTful API for smart mattress monitoring system supporting patients, doctors, customers, and admins.",
        "contact": {
            "name": "SmartRest Support",
            "email": "support@smartrest.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "SmartRest API Server"
        }
    ],
    "paths": {
...
        "/sensors/data": {
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Store sensor readings",
                "description": "Upload batched sensor readings from smart bed",
                "operationId": "storeSensorData",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "bed_id",
                                    "readings"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                    },
                                    "bed_id": {
                                        "type": "string",
                                        "example": "BED-12345"
                                    },
                                    "readings": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "sensor_type",
                                                "sensor_value"
                                            ],
                                            "properties": {
                                                "sensor_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pressure",
                                                        "heart_rate",
                                                        "breathing_rate",
                                                        "temperature",
                                                        "humidity",
                                                        "body_movement",
                                                        "posture",
                                                        "vibration",
                                                        "sleep_apnea"
                                                    ],
                                                    "example": "heart_rate"
                                                },
                                                "sensor_value": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 72.5
                                                },
                                                "sensor_unit": {
                                                    "type": "string",
                                                    "example": "bpm"
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-24T12:00:00Z"
                                                },
                                                "additional_metadata": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sensor data stored successfully"
                                        },
                                        "readings_count": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sensors/latest": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get latest sensor readings",
                "description": "Get latest readings for each sensor type for the patient",
                "operationId": "getLatestSensorData",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID (required for doctors, not for patients)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest sensor readings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "patient_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "readings": {
                                            "properties": {
                                                "heart_rate": {
                                                    "properties": {
                                                        "value": {
                                                            "type": "number",
                                                            "example": 72.5
                                                        },
                                                        "unit": {
                                                            "type": "string",
                                                            "example": "bpm"
                                                        },
                                                        "timestamp": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "breathing_rate": {
                                                    "properties": {
                                                        "value": {
                                                            "type": "number",
                                                            "example": 16.8
                                                        },
                                                        "unit": {
                                                            "type": "string",
                                                            "example": "breaths/min"
                                                        },
                                                        "timestamp": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "temperature": {
                                                    "properties": {
                                                        "value": {
                                                            "type": "number",
                                                            "example": 36.7
                                                        },
                                                        "unit": {
                                                            "type": "string",
                                                            "example": "°C"
                                                        },
                                                        "timestamp": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - No access to this patient",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have access to this patient"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}